+++ cw-orch-daemon/tests/instantiate2.rs
@@ -15,7 +15,10 @@ pub mod test {
     #[test]
     #[serial_test::serial]
     fn instantiate2() -> anyhow::Result<()> {
-        let app = Daemon::builder(networks::LOCAL_JUNO).build().unwrap();
+        let app = Daemon::builder(networks::LOCAL_JUNO)
+            .is_test(true)
+            .build()
+            .unwrap();
 
         let salt = Binary(vec![12, 89, 156, 63]);
         let mock_contract = MockContract::new("mock-contract", app.clone());